Fixes, language model
