+++ tests/test_multi_stage.py
@@ -5,7 +5,6 @@ import pytest
 from fast_llm.engine.distributed.distributed import Distributed
 from fast_llm.engine.multi_stage.config import FastLLMModelConfig
 from fast_llm.engine.multi_stage.fast_llm_model import FastLLMModel
-from fast_llm.layers.decoder.block import DecoderBlock
 from fast_llm.utils import Assert
 from tests.utils.dataset import get_model_test_dataset
 from tests.utils.model_configs import ModelTestingGroup
@@ -42,8 +41,7 @@ def test_frozen_weights(model_testing_config):
         model_frozen._num_stages,
     )
     frozen_parameter_counts = [
-        sum(p.numel() for p in layer.unwrap().mlp.parameters()) if isinstance(layer.unwrap(), DecoderBlock) else 0
-        for layer in model_ref.base_model.get_layers()
+        sum(p.numel() for p in layer.mlp.parameters()) for layer in model_ref.base_model.decoder.layers
     ]
 
     # Make sure each layer has its own buffer so the check below works.
